change uint32_t to uint64_t

+++ sha1/sha1.h
@@ -13,6 +13,6 @@ typedef struct {
 void sha1Init(sha1Context *);
 void sha1ProcessBlock(sha1Context *);
 void sha1PadMessage(sha1Context *);
-void sha1Input(uint8_t *, uint32_t messageLength, sha1Context *);
+void sha1Input(uint8_t *, uint64_t messageLength, sha1Context *);
 void sha1Output(sha1Context *);
